+++ APOD/APOD/Infrastructure/Network/Common/NetworkError.swift
