+++ APOD/APOD/View/Controller/ViewController.swift
@@ -309,7 +309,7 @@ class ViewController: UIViewController, WKNavigationDelegate {
         /// ref. https://developer.apple.com/library/archive/documentation/Performance/Conceptual/EnergyGuide-iOS/PrioritizeWorkWithQoS.html
         /// ref. https://unnnyong.com/2020/05/14/ios-thread-queue-gcd-qos/
         DispatchQueue.global(qos: .userInteractive).async {
-            let endpoint: Endpoint<Apod> = APIEndpoints.getApod(with: RequestDTO())
+            let endpoint: Endpoint<Apod> = APIEndpoints.getApod(with: ApodDTO())
             
             APIProvider.shared.request(with: endpoint) { [weak self] result in
                 /// `[weak self]`로 fetchApod()의 escaping closure (completion)가 ViewController를 약하게 참조 (Memory Leaks 방지)
